requirements.txt
flask==1.1.2
flask-login==0.5.0
flask-migrate==2.7.0
flask-wtf==0.14.3
flask-sqlalchemy==2.4.4
sqlalchemy==1.3.23
python-decouple==3.4
gunicorn==20.0.4
jinja2==2.11.3
psycopg2==2.9.1
git+git://github.com/spykard/populartimes.git@ef1a5331f20eaa7616d919841ee77fc2ed157091
